Remove promote playbook from opendev-promote-docker-image
This job inherits from promote-docker-image in zuul-jobs which
contains this playbook already; therefore, this job simply
duplicated the playbook.  Remove it.

(You may note this job doesn't add anything to the parent, it
exists only for naming consistency, so that people can just use
the trio of opendev jobs instead of the upstream ones.)

Change-Id: Ib3afc56eadcffe889713dab49f12d836c87c0740