Only use ipv6 dns if nodepool uses ipv6
We think we have growing pains in some clouds where ipv6 is available
but we want to force the instances to use ipv4 dns anyway. We can do
this by telling nodepool to use ipv4 for the instance ip which we can
then check in our role to configure unbound.

Do this to rule out ipv6 as a problem in these clouds.

Change-Id: I354015df64032e70422231f1105807f024fe2393