Open port 5001 for the registry proxy
Previously we had opened port 5000 for the buildset registry. But we
also run a buildset registry proxy on port 5001. Because docker manages
the ipv4 firewall rules itself it was opening ipv4 5001 but ignoring
ipv6 5001. This meant that jobs that use ipv6 addrs as primary addrs
were failing to talk ot the registry proxy.

Fix this by opening the range 5000:5001.

Change-Id: I88abcd7fd99fb7dcc7726ef9a52f9b022729f9d5