Test removal of fetch-zuul-cloner
During the original import of Zuul v2 jobs to become legacy jobs in
v3, we included fetch-zuul-cloner in the base job. Soon thereafter
we added a legacy-base job to act as an abstraction layer, and added
fetch-zuul-cloner to it as well, but never finished removing it from
the base job. A more detailed analysis[*] of the history can be
found on the openstack-discuss mailing list, but suffice to say
we're well past due to finish the transition and it's causing pains
for users of modern v3-native jobs which don't need the old
compatibility shim.

[*] http://lists.openstack.org/pipermail/openstack-discuss/2019-June/006881.html

Change-Id: I3b8d038fa05de084fd23058e323335503c3f793b