Add base-minimal job
This is required by zuul-jobs for its testing.  We could put it
only in the Zuul tenant, but we would need to duplicate the secret,
so this is lazier.  It shouldn't hurt to have it in all the tenants.
We can set allowed-projects so that only zuul-jobs can use it in
order to avoid accidental use and confusion.

Change-Id: I8966737c580628b4fa097e32594c8a4e99926147