Shard build logs with build uuid in all base jobs
Use the first 3 characters of the build uuid to shard log paths and
container names. This will ensure we distribute objects across
containers more evently. THis is particularly useful for ceph where
having many objects in a single container makes it sad.

Note this also changes the container name prefix to be opendev specific
to avoid name collisions on Ceph clusters as container names must be
globally unique on ceph.

Change-Id: Iac6cd3ead276ba514ce8e8e31e202b0a61b12024