Add cleanup phase to base(-test)
This is the first change to add a cleanup phase to our jobs. This
cleanup phase will use the raw module to get disk and networking info
from hosts. This info can be useful if the tests fail due to
"networking" problems when full ansible modules can't be bootstrapped.

I tested this usage of the raw module against a host with a full root
disk so it should work just fine.

Adding to base-test to test it in general usage before adding to base
and base-minimal.

Change-Id: I0c1569d3ca699cbd6be035f987e90805863eb6b2