Add cleanup playbook to all base jobs
This cleanup playbook is used as a last effort to grab debugging info
from hosts that might othewise be missed if we run out of disk space or
have other problems. Using the raw module we ssh in and get df output as
well as ip link and ip addr output for all jobs in the cleanup phase.

Because cleanup runs after the post playbook this info is only captured
in the executor logs.

Depends-On: https://review.opendev.org/#/c/681552/
Change-Id: Ib5d2cb1dab4da8a9a6d0f3e1721fff395b564044