Timeout cleanup playbook tasks
This adds timeouts to cleanup playbook tasks so that if they have
problems running the entire job does not stall. We noticed that df in
particular can hang on broken nfs mounts and if that happens the job's
cleanup playbook does not end (even after a couple days).

This change adds the timeouts to the testing playbook. If testing shows
this works well we will add it to the production playbooks too.

Change-Id: Ibc875bf99e6da29e2fffbddae8590ecde06b5c3b