Add number of zuul job attempts to logstash events
This will help us better track how often jobs are being rerun. We should
be able to query for numeric values of this field and count the number
of jobs that had more than one attempt. From that we should also be able
to identify if specific jobs hit these problems more than others.

Change-Id: I6ad21ad3a55140356b9a5f19e5e5bb83663898f3