Cleanup workaround for importlib-resources bad version
Upstream of us importlib-resources has made a couple releases that fix
the problems that were in 1.2.0 with python2.7 and python3.5. I've tried
to confirm this locally installing virtualenv and making virtualenvs
under python2.7 and python3.5 and this works with new libs.

Since those libs have released we've built and uploaded new test node
images. This means we can remove our workaround now.

Note that we want importlib-resources 1.3.1 to avoid the bugs in the
earlier release.

Change-Id: I70d1161b8639664fcecaddaec6a7cb3a0aac43ec