Remove env var for tarballs mirror
This host has migrated to tarballs.opendev.org, but rather than
updating this caching rule, an audit of codesearch and a cursory look
at several mirror logs has determined this wasn't ever really used.

Since it is now on AFS, we could export this directly from the
filesystem on mirror nodes and implicitly use AFS caching if we need
to reinstate this.

Change-Id: I4feb2a2229dcf0626ec8ee8494ad6801468c1440
Story: #2006598
Task: #39014