virtualenv-config : add download for virtualenv seeder
As described in the inline comments, this is working around a broken
version of setuptools vendored by virtualenv currently.  Specify in a
config file that it should download the latest helper tools, rather
then rely on the inbuilt versions.

Add to the base-test pre-playbook; after testing we can move to the
base playbook.

Change-Id: Ib17017637eae81a3ff57302e7c77945f2045b5ac