base: add bulk log download script
This adds a bulk download script using the role added with
Ic33732adbfd3210191bf4976c3ee316cfc50568e.

An artifact is added to the build results with a quick copy-paste
command-line which will grab and run the script.  Or it can be
downloaded an run manually.

This has been tested with prior commits to base-test.

Change-Id: Ib6f81873368261f7ddb48aa9e23906d8045219d3